+++ programs/drift/src/lib.rs
@@ -1560,6 +1560,21 @@ pub mod drift {
         handle_initialize_pyth_pull_oracle(ctx, feed_id)
     }
 
+    pub fn initialize_pyth_lazer_oracle(
+        ctx: Context<InitPythLazerOracle>,
+        feed_id: u32,
+    ) -> Result<()> {
+        handle_initialize_pyth_lazer_oracle(ctx, feed_id)
+    }
+
+    pub fn post_pyth_lazer_oracle_update(
+        ctx: Context<UpdatePythLazerOracle>,
+        feed_id: u32,
+        pyth_message: Vec<u8>,
+    ) -> Result<()> {
+        handle_update_pyth_lazer_oracle(ctx, feed_id, pyth_message)
+    }
+
     pub fn initialize_high_leverage_mode_config(
         ctx: Context<InitializeHighLeverageModeConfig>,
         max_users: u32,