+++ programs/drift/src/state/oracle.rs
@@ -12,6 +12,7 @@ use crate::error::ErrorCode::{InvalidOracle, UnableToLoadOracle};
 use crate::math::safe_unwrap::SafeUnwrap;
 use crate::state::load_ref::load_ref;
 use crate::state::perp_market::PerpMarket;
+use crate::state::pyth_lazer_oracle::PythLazerOracle;
 use crate::state::traits::Size;
 use crate::validate;
 
@@ -124,6 +125,44 @@ pub enum OracleSource {
     Pyth1MPull,
     PythStableCoinPull,
     SwitchboardOnDemand,
+    PythLazer,
+}
+
+impl OracleSource {
+    pub fn is_pyth_pull_oracle(&self) -> bool {
+        matches!(
+            self,
+            OracleSource::PythPull
+                | OracleSource::Pyth1KPull
+                | OracleSource::Pyth1MPull
+                | OracleSource::PythStableCoinPull
+        )
+    }
+
+    pub fn is_pyth_push_oracle(&self) -> bool {
+        matches!(
+            self,
+            OracleSource::Pyth
+                | OracleSource::Pyth1K
+                | OracleSource::Pyth1M
+                | OracleSource::PythStableCoin
+        )
+    }
+
+    pub fn get_pyth_multiple(&self) -> u128 {
+        match self {
+            OracleSource::Pyth
+            | OracleSource::PythPull
+            | OracleSource::PythLazer
+            | OracleSource::PythStableCoin
+            | OracleSource::PythStableCoinPull => 1,
+            OracleSource::Pyth1K | OracleSource::Pyth1KPull => 1000,
+            OracleSource::Pyth1M | OracleSource::Pyth1MPull => 1000000,
+            _ => {
+                panic!("Calling get_pyth_multiple on non-pyth oracle source");
+            }
+        }
+    }
 }
 
 #[derive(Default, Clone, Copy, Debug)]
@@ -151,10 +190,12 @@ pub fn get_oracle_price(
     clock_slot: u64,
 ) -> DriftResult<OraclePriceData> {
     match oracle_source {
-        OracleSource::Pyth => get_pyth_price(price_oracle, clock_slot, 1, false),
-        OracleSource::Pyth1K => get_pyth_price(price_oracle, clock_slot, 1000, false),
-        OracleSource::Pyth1M => get_pyth_price(price_oracle, clock_slot, 1000000, false),
-        OracleSource::PythStableCoin => get_pyth_stable_coin_price(price_oracle, clock_slot, false),
+        OracleSource::Pyth => get_pyth_price(price_oracle, clock_slot, oracle_source),
+        OracleSource::Pyth1K => get_pyth_price(price_oracle, clock_slot, oracle_source),
+        OracleSource::Pyth1M => get_pyth_price(price_oracle, clock_slot, oracle_source),
+        OracleSource::PythStableCoin => {
+            get_pyth_stable_coin_price(price_oracle, clock_slot, oracle_source)
+        }
         OracleSource::Switchboard => get_switchboard_price(price_oracle, clock_slot),
         OracleSource::SwitchboardOnDemand => get_sb_on_demand_price(price_oracle, clock_slot),
         OracleSource::QuoteAsset => Ok(OraclePriceData {
@@ -164,21 +205,22 @@ pub fn get_oracle_price(
             has_sufficient_number_of_data_points: true,
         }),
         OracleSource::Prelaunch => get_prelaunch_price(price_oracle, clock_slot),
-        OracleSource::PythPull => get_pyth_price(price_oracle, clock_slot, 1, true),
-        OracleSource::Pyth1KPull => get_pyth_price(price_oracle, clock_slot, 1000, true),
-        OracleSource::Pyth1MPull => get_pyth_price(price_oracle, clock_slot, 1000000, true),
+        OracleSource::PythPull => get_pyth_price(price_oracle, clock_slot, oracle_source),
+        OracleSource::Pyth1KPull => get_pyth_price(price_oracle, clock_slot, oracle_source),
+        OracleSource::Pyth1MPull => get_pyth_price(price_oracle, clock_slot, oracle_source),
         OracleSource::PythStableCoinPull => {
-            get_pyth_stable_coin_price(price_oracle, clock_slot, true)
+            get_pyth_stable_coin_price(price_oracle, clock_slot, oracle_source)
         }
+        OracleSource::PythLazer => get_pyth_price(price_oracle, clock_slot, oracle_source),
     }
 }
 
 pub fn get_pyth_price(
     price_oracle: &AccountInfo,
     clock_slot: u64,
-    multiple: u128,
-    is_pull_oracle: bool,
+    oracle_source: &OracleSource,
 ) -> DriftResult<OraclePriceData> {
+    let multiple = oracle_source.get_pyth_multiple();
     let mut pyth_price_data: &[u8] = &price_oracle
         .try_borrow_data()
         .or(Err(crate::error::ErrorCode::UnableToLoadOracle))?;
@@ -189,7 +231,7 @@ pub fn get_pyth_price(
     let mut oracle_precision: u128;
     let published_slot: u64;
 
-    if is_pull_oracle {
+    if oracle_source.is_pyth_pull_oracle() {
         let price_message = pyth_solana_receiver_sdk::price_update::PriceUpdateV2::try_deserialize(
             &mut pyth_price_data,
         )
@@ -198,7 +240,7 @@ pub fn get_pyth_price(
         oracle_conf = price_message.price_message.conf;
         oracle_precision = 10_u128.pow(price_message.price_message.exponent.unsigned_abs());
         published_slot = price_message.posted_slot;
-    } else {
+    } else if oracle_source.is_pyth_push_oracle() {
         let price_data = pyth_client::cast::<pyth_client::Price>(pyth_price_data);
         oracle_price = price_data.agg.price;
         oracle_conf = price_data.agg.conf;
@@ -216,6 +258,12 @@ pub fn get_pyth_price(
 
         oracle_precision = 10_u128.pow(price_data.expo.unsigned_abs());
         published_slot = price_data.valid_slot;
+    } else {
+        let price_data = PythLazerOracle::try_deserialize(&mut pyth_price_data).unwrap();
+        oracle_price = price_data.price;
+        oracle_conf = price_data.conf;
+        oracle_precision = 10_u128.pow(price_data.exponent.unsigned_abs());
+        published_slot = price_data.posted_slot;
     }
 
     if oracle_precision <= multiple {
@@ -258,9 +306,9 @@ pub fn get_pyth_price(
 pub fn get_pyth_stable_coin_price(
     price_oracle: &AccountInfo,
     clock_slot: u64,
-    is_pull_oracle: bool,
+    oracle_source: &OracleSource,
 ) -> DriftResult<OraclePriceData> {
-    let mut oracle_price_data = get_pyth_price(price_oracle, clock_slot, 1, is_pull_oracle)?;
+    let mut oracle_price_data = get_pyth_price(price_oracle, clock_slot, oracle_source)?;
 
     let price = oracle_price_data.price;
     let confidence = oracle_price_data.confidence;