+++ CHANGELOG.md
@@ -9,9 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
-- programL add ix to log user balances ([#1366](https://github.com/drift-labs/protocol-v2/pull/1366))
+- program: pyth lazer integration ([#1361](https://github.com/drift-labs/protocol-v2/pull/1361))
+- program: add ix to log user balances ([#1366](https://github.com/drift-labs/protocol-v2/pull/1366))
 
 ### Fixes
+
 program: fix force delete user for token 2022 ([#1358](https://github.com/drift-labs/protocol-v2/pull/1358))
 
 ### Breaking